requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
python-dotenv==1.0.1
